Solving issues in testing

+++ ParetoLib/Search/ResultSet.py
@@ -462,7 +462,7 @@ class ResultSet:
         if sec > 0.0 and not blocking:
             plt.ion()
             plt.show()
-            plt.pause(int(sec))
+            plt.pause(float(sec))
         else:
             plt.ioff()
             plt.show()
@@ -535,7 +535,7 @@ class ResultSet:
         if sec > 0.0 and not blocking:
             plt.ion()
             plt.show()
-            plt.pause(int(sec))
+            plt.pause(float(sec))
         else:
             plt.ioff()
             plt.show()
@@ -602,7 +602,7 @@ class ResultSet:
         if sec > 0.0 and not blocking:
             plt.ion()
             plt.show()
-            plt.pause(int(sec))
+            plt.pause(float(sec))
         else:
             plt.ioff()
             plt.show()
@@ -699,7 +699,7 @@ class ResultSet:
         if sec > 0.0 and not blocking:
             plt.ion()
             plt.show()
-            plt.pause(int(sec))
+            plt.pause(float(sec))
         else:
             plt.ioff()
             plt.show()
